Update and rename swith2.py to switсh2.py

+++ modules/switсh2.py
@@ -7,12 +7,17 @@ def switch(client, message):
     text = ' '.join(message.command[1:])
     ru_keys = """ёйцукенгшщзхъфывапролджэячсмитьбю.Ё"№;%:?ЙЦУКЕНГШЩЗХЪФЫВАПРОЛДЖЭЯЧСМИТЬБЮ,"""
     en_keys = """eicykengшщzh_fiwaproldgeiчsmit_bu.E"№;%:?ICYKENGШЩZH_FIWAPROLDGEIЧSMIT_BU,"""
-    if message.reply_to_message:
-        reply_text = message.reply_to_message.text
-        message.delete()
-        change = str.maketrans(ru_keys + en_keys, en_keys + ru_keys)
-        reply_text = str.translate(reply_text, change)
-        client.send_message(message.chat.id, reply_text)
+    if text == '':
+        if message.reply_to_message:
+            reply_text = message.reply_to_message.text
+            message.delete()
+            change = str.maketrans(ru_keys + en_keys, en_keys + ru_keys)
+            reply_text = str.translate(reply_text, change)
+            client.send_message(message.chat.id, reply_text)
+        else:
+            message.edit('No text for switch')
+            time.sleep(3)
+            message.delete()
     else:
         message.delete()
         change = str.maketrans(ru_keys + en_keys, en_keys + ru_keys)
